Fixed ring buffer size limited to 255 elements, instead of the intended 65535 elements.

+++ LUFA/Drivers/Misc/RingBuffer.h
@@ -113,7 +113,7 @@
 			uint8_t* Out; /**< Current retrieval location in the circular buffer. */
 			uint8_t* Start; /**< Pointer to the start of the buffer's underlying storage array. */
 			uint8_t* End; /**< Pointer to the end of the buffer's underlying storage array. */
-			uint8_t  Size; /**< Size of the buffer's underlying storage array. */
+			uint16_t Size; /**< Size of the buffer's underlying storage array. */
 			uint16_t Count; /**< Number of bytes currently stored in the buffer. */
 		} RingBuffer_t;
